Move fetchSandboxes to utils
