+++ src/view/dataElements/identityMap/utils/namespacesUtils.js
@@ -10,7 +10,7 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import fetchSandboxes from "../../xdmObject/helpers/fetchSandboxes";
+import fetchSandboxes from "../../../utils/fetchSandboxes";
 import fetchNamespaces from "./fetchNamespaces";
 
 const isNotECID = namespace => {
