+++ src/view/dataElements/xdmObject.jsx
@@ -25,7 +25,7 @@ import useChanged from "../utils/useChanged";
 import FormikPicker from "../components/formikReactSpectrum3/formikPicker";
 import FormikPagedComboBox from "../components/formikReactSpectrum3/formikPagedComboBox";
 import DEFAULT_SANDBOX_NAME from "./xdmObject/constants/defaultSandboxName";
-import fetchSandboxes from "./xdmObject/helpers/fetchSandboxes";
+import fetchSandboxes from "../utils/fetchSandboxes";
 import fetchSchemasMeta from "./xdmObject/helpers/fetchSchemasMeta";
 import fetchSchema from "./xdmObject/helpers/fetchSchema";
